✨ Add environment configuration, improve asset loading, and enhance error handling in the App component
