+++ packages/frontend/.env-example
@@ -1,0 +1,2 @@
+VITE_API_BASE_URL=http://127.0.0.1:8000
+NODE_ENV=development
