+++ packages/frontend/.gitignore
@@ -23,3 +23,9 @@ dist-ssr
 *.sln
 *.sw?
 /stats.html
+
+
+.env
+.env.local
+
+.vscode
