+++ packages/frontend/package.json
@@ -16,19 +16,19 @@
     "prepare": "husky"
   },
   "dependencies": {
-    "@mantine/carousel": "^7.17.0",
-    "@mantine/charts": "^7.17.0",
-    "@mantine/code-highlight": "^7.17.0",
-    "@mantine/core": "^7.17.0",
-    "@mantine/dates": "^7.17.0",
-    "@mantine/dropzone": "^7.17.0",
-    "@mantine/form": "^7.17.0",
-    "@mantine/hooks": "^7.17.0",
-    "@mantine/modals": "^7.17.0",
-    "@mantine/notifications": "^7.17.0",
-    "@mantine/nprogress": "^7.17.0",
-    "@mantine/spotlight": "^7.17.0",
-    "@mantine/tiptap": "^7.17.0",
+    "@mantine/carousel": "^7.17.1",
+    "@mantine/charts": "^7.17.1",
+    "@mantine/code-highlight": "^7.17.1",
+    "@mantine/core": "^7.17.1",
+    "@mantine/dates": "^7.17.1",
+    "@mantine/dropzone": "^7.17.1",
+    "@mantine/form": "^7.17.1",
+    "@mantine/hooks": "^7.17.1",
+    "@mantine/modals": "^7.17.1",
+    "@mantine/notifications": "^7.17.1",
+    "@mantine/nprogress": "^7.17.1",
+    "@mantine/spotlight": "^7.17.1",
+    "@mantine/tiptap": "^7.17.1",
     "@pywebflow/api": "workspace:*",
     "@radix-ui/react-dialog": "^1.1.6",
     "@radix-ui/react-navigation-menu": "^1.2.5",
@@ -37,22 +37,24 @@
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/react-toast": "^1.2.6",
     "@radix-ui/react-tooltip": "^1.1.8",
-    "@radix-ui/themes": "^3.2.0",
+    "@radix-ui/themes": "^3.2.1",
     "@tabler/icons-react": "^3.30.0",
-    "@tailwindcss/postcss": "^4.0.7",
-    "@tailwindcss/vite": "^4.0.7",
+    "@tailwindcss/postcss": "^4.0.9",
+    "@tailwindcss/vite": "^4.0.9",
     "@tiptap/extension-link": "^2.11.5",
     "@tiptap/pm": "^2.11.5",
     "@tiptap/react": "^2.11.5",
     "@tiptap/starter-kit": "^2.11.5",
-    "@xyflow/react": "^12.4.3",
+    "@xyflow/react": "^12.4.4",
     "autoprefixer": "^10.4.20",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.13",
+    "dompurify": "^3.2.4",
     "embla-carousel-react": "^8.5.2",
     "framer-motion": "^12.4.7",
-    "lucide-react": "^0.475.0",
+    "html-react-parser": "^5.2.2",
+    "lucide-react": "^0.477.0",
     "next-themes": "^0.4.4",
     "postcss": "^8.5.3",
     "react": "^19.0.0",
@@ -62,30 +64,31 @@
     "react-remark": "^2.1.0",
     "react-router-dom": "^7.2.0",
     "recharts": "^2.15.1",
-    "sass": "^1.85.0",
-    "tailwind-merge": "^3.0.1",
-    "tailwindcss": "^4.0.7",
+    "sass": "^1.85.1",
+    "tailwind-merge": "^3.0.2",
+    "tailwindcss": "^4.0.9",
     "tailwindcss-animate": "^1.0.7"
   },
   "license": "apache-2.0",
   "devDependencies": {
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-prettier": "^5.2.3",
-    "@types/node": "^22.13.4",
+    "@types/node": "^22.13.8",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
-    "@typescript-eslint/eslint-plugin": "^8.24.1",
-    "@typescript-eslint/parser": "^8.24.1",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
+    "@typescript-eslint/parser": "^8.25.0",
     "@vitejs/plugin-react": "^4.3.4",
-    "eslint": "^9.20.1",
-    "eslint-plugin-react-hooks": "^5.1.0",
+    "@vitejs/plugin-react-swc": "^3.8.0",
+    "eslint": "^9.21.0",
+    "eslint-config-prettier": "^10.0.2",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "husky": "^9.1.7",
     "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
     "rollup-plugin-visualizer": "^5.14.0",
-    "typescript": "^5.7.3",
-    "vite": "^6.1.1"
+    "typescript": "^5.8.2",
+    "vite": "^6.2.0"
   },
   "keywords": [
     "pywebflow",
