+++ packages/frontend/src/App.tsx
@@ -12,20 +12,30 @@ import {
 } from '@pywebflow/api/src/sidebar.ts';
 
 export function App() {
-  const [sidebarItems, setSidebarItems] = useState<SidebarResponse['items']>(
-    [],
-  );
+  const [sidebarItems, setSidebarItems] = useState<SidebarResponse['items']>([]);
   const [sidebarVisible, setSidebarVisible] = useState<boolean>(false);
   const [sidebarLabel, setSidebarLabel] = useState<string>('');
   const [defaultOpen, setDefaultOpen] = useState<boolean>(false);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
 
   useEffect(() => {
     const fetchSidebarItems = async () => {
-      const response: SidebarResponse = await getSidebarItems();
-      setSidebarItems(response.items);
-      setSidebarVisible(response.visible);
-      setSidebarLabel(response.label);
-      setDefaultOpen(response.default_open);
+      try {
+        setLoading(true);
+        setError(null);
+
+        const response: SidebarResponse = await getSidebarItems();
+        setSidebarItems(response.items);
+        setSidebarVisible(response.visible);
+        setSidebarLabel(response.label);
+        setDefaultOpen(response.default_open);
+      } catch (err) {
+        console.error('Error fetching sidebar items:', err);
+        setError('Failed to load sidebar');
+      } finally {
+        setLoading(false);
+      }
     };
 
     fetchSidebarItems();
@@ -35,16 +45,20 @@ export function App() {
     <div className="App h-screen overflow-hidden">
       <SidebarProvider defaultOpen={defaultOpen}>
         <div className="flex h-full">
-          {sidebarVisible && (
+          {loading && <p className="p-4 text-gray-500">Loading sidebar...</p>}
+          {error && <p className="p-4 text-red-500">{error}</p>}
+
+          {!loading && !error && sidebarVisible && (
             <AppSidebar items={sidebarItems} label={sidebarLabel} />
           )}
-          {sidebarVisible && <SidebarTrigger />}
+          {!loading && !error && sidebarVisible && <SidebarTrigger />}
+
           <div className="flex-1 overflow-auto">
             <Status />
             <ControlsComp />
             <MinimapComp />
             <BackgroundWrapper />
-            <InjectedHtml /> {/* Add the InjectedHtml component here */}
+            <InjectedHtml />
           </div>
         </div>
       </SidebarProvider>
