+++ packages/frontend/src/Layout.tsx
@@ -30,21 +30,29 @@ export default function Layout() {
   const { theme, setTheme, systemTheme } = useTheme();
   const currentTheme = theme === 'system' ? systemTheme : theme;
 
-  useEffect(() => {
-    const loadInitialConfig = async () => {
-      try {
-        const fetchedConfig = await getConfig();
-        if (fetchedConfig[0].colorMode) {
+useEffect(() => {
+  const loadInitialConfig = async () => {
+    try {
+      const fetchedConfig = await getConfig();
+
+      if (Array.isArray(fetchedConfig) && fetchedConfig.length > 0) {
+        if (fetchedConfig[0]?.colorMode) {
           setTheme(fetchedConfig[0].colorMode);
         }
         setConfig(fetchedConfig[0]);
-      } catch (error) {
-        console.error('Error fetching initial config:', error);
+      } else {
+        setConfig({});
       }
-    };
+    } catch (error) {
+      console.error("Error fetching initial config:", error);
+      setConfig({}); // Fallback to empty object on error
+    }
+  };
 
-    loadInitialConfig();
-  }, [setTheme]);
+  loadInitialConfig();
+}, [setTheme]);
+
+
 
   useEffect(() => {
     const updateProgress = () => {
