+++ packages/frontend/src/Root.tsx
@@ -1,3 +1,4 @@
+import './styles.ts';
 import React, { Suspense, useEffect, useState } from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
@@ -5,18 +6,11 @@ import { MantineProvider } from '@mantine/core';
 import { ThemeProvider } from 'next-themes';
 import { Theme } from '@radix-ui/themes';
 import { HelmetProvider } from 'react-helmet-async';
-import './styles/index.scss';
-import '@xyflow/react/dist/style.css';
-import '@radix-ui/themes/styles.css';
-import '@mantine/core/styles.css';
-import { loadAndInjectAssets } from './hooks/assets.ts';
 import { getConfig } from '@pywebflow/api/src/config';
 
 // Lazy load components with preloading
 const Layout = React.lazy(() => import(/* webpackPreload: true */ './Layout'));
-const MetaData = React.lazy(
-  () => import(/* webpackPreload: true */ './components/Metadata'),
-);
+const MetaData = React.lazy(() => import(/* webpackPreload: true */ './components/Metadata'));
 
 // Preload critical components asynchronously
 const preloadAssets = () => {
@@ -28,27 +22,20 @@ preloadAssets();
 
 const Root = () => {
   const [mounted, setMounted] = useState(false);
-  const [defaultTheme, setDefaultTheme] = useState<'light' | 'dark' | 'system'>(
-    'system',
-  );
+  const [defaultTheme, setDefaultTheme] = useState<'light' | 'dark' | 'system'>('system');
 
   useEffect(() => {
-    // Load assets dynamically
-    loadAndInjectAssets();
-
     // Fetch config data
     const fetchConfig = async () => {
       try {
         const fetchedConfig = await getConfig();
-        const config = fetchedConfig.data[0];
-        if (config.colorMode) {
-          setDefaultTheme(config.colorMode);
+        if (fetchedConfig?.data?.length > 0) {
+          setDefaultTheme(fetchedConfig.data[0].colorMode || 'system');
         }
       } catch (error) {
         console.error('Error fetching config:', error);
       }
     };
-
     fetchConfig();
   }, []);
 
@@ -58,12 +45,7 @@ const Root = () => {
 
   return (
     <HelmetProvider>
-      <ThemeProvider
-        attribute="class"
-        defaultTheme={defaultTheme}
-        enableSystem
-        disableTransitionOnChange
-      >
+      <ThemeProvider attribute="class" defaultTheme={defaultTheme} enableSystem disableTransitionOnChange>
         <MantineProvider>
           <Theme>
             {mounted && (
