+++ packages/frontend/src/components/Metadata.tsx
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import getMetadata, { Metadata } from '@pywebflow/api/src/metadata.ts';
+import { injectAssets } from '../hooks/assets.ts';
+
 const MetaData: React.FC = () => {
   const [metadata, setMetadata] = useState<Metadata>({
     title: 'PyWebflow',
@@ -8,7 +10,9 @@ const MetaData: React.FC = () => {
   });
 
   useEffect(() => {
-    const fetchMetadata = async () => {
+    const loadMetadataAndAssets = async () => {
+      await injectAssets(); // Load CSS & JS before fetching metadata
+
       try {
         const data = await getMetadata();
         setMetadata((prevMetadata) => ({
@@ -20,35 +24,23 @@ const MetaData: React.FC = () => {
       }
     };
 
-    fetchMetadata();
+    loadMetadataAndAssets();
   }, []);
 
   return (
     <Helmet>
       <title>{metadata.title}</title>
-      {metadata.description && (
-        <meta name="description" content={metadata.description} />
-      )}
-      {metadata.keywords && (
-        <meta name="keywords" content={metadata.keywords} />
-      )}
+      {metadata.description && <meta name="description" content={metadata.description} />}
+      {metadata.keywords && <meta name="keywords" content={metadata.keywords} />}
       {metadata.author && <meta name="author" content={metadata.author} />}
-      {metadata.viewport && (
-        <meta name="viewport" content={metadata.viewport} />
-      )}
+      {metadata.viewport && <meta name="viewport" content={metadata.viewport} />}
       {metadata.charset && <meta charSet={metadata.charset} />}
       {metadata.robots && <meta name="robots" content={metadata.robots} />}
       {metadata.canonical && <link rel="canonical" href={metadata.canonical} />}
-      {metadata.ogTitle && (
-        <meta property="og:title" content={metadata.ogTitle} />
-      )}
-      {metadata.ogDescription && (
-        <meta property="og:description" content={metadata.ogDescription} />
-      )}
+      {metadata.ogTitle && <meta property="og:title" content={metadata.ogTitle} />}
+      {metadata.ogDescription && <meta property="og:description" content={metadata.ogDescription} />}
       {metadata.ogUrl && <meta property="og:url" content={metadata.ogUrl} />}
-      {metadata.ogImage && (
-        <meta property="og:image" content={metadata.ogImage} />
-      )}
+      {metadata.ogImage && <meta property="og:image" content={metadata.ogImage} />}
     </Helmet>
   );
 };
