+++ packages/frontend/src/styles.ts
@@ -1,0 +1,4 @@
+import './styles/index.scss';
+import '@xyflow/react/dist/style.css';
+import '@radix-ui/themes/styles.css';
+import '@mantine/core/styles.css';
