+++ webflow/modules/__init__.py
@@ -1,4 +1,3 @@
 from webflow.modules.arguments import parse_arguments
 from webflow.modules.mount import mount_static_files
 from webflow.modules.routes import WebFlow_API
-from webflow.modules.serve import app
